Rename files and fix indentation
