--- src/TrillekGame.cpp
@@ -1,9 +1,0 @@
-#include "TrillekGame.h"
-
-namespace trillek {
-    TrillekScheduler TrillekGame::scheduler;
-    FakeSystem TrillekGame::fake_system;
-    OS TrillekGame::glfw_os;
-    graphics::System TrillekGame::gl_sys;
-    bool TrillekGame::close_window = false;
-}
